refactor(tests): introduce integration_tests
Make the repo structure more understable w.r.t testing.
