+++ src/http/mod.rs
@@ -31,6 +31,7 @@ mod transport;
 #[cfg(test)]
 mod tests;
 
+pub use crate::types::http::HttpConfig;
 pub use raw::RawServer as HttpRawServer;
 pub use raw::RawServerEvent as HttpRawServerEvent;
 pub use raw::TypedResponder as HttpTypedResponder;
